Refactor beep generation code to fix incorrect frequencies and other issues.
Fixes #5023.
Authors: Austin Hicks <camlorn38@gmail.com>, Davy Kager <mail@davykager.nl>

+++ source/tones.py
@@ -29,7 +29,7 @@ def beep(hz,length,left=50,right=50):
 	@param left: volume of the left channel (0 to 100)
 	@type left: integer
 	@param right: volume of the right channel (0 to 100)
-	@type right: float
+	@type right: integer
 	""" 
 	log.io("Beep at pitch %s, for %s ms, left volume %s, right volume %s"%(hz,length,left,right))
 	if not player:
